services/web/requirements.txt
Flask==3.0.0
gunicorn==21.2.0
google-cloud-storage==2.18.0
google-auth==2.35.0
requests>=2.31.0   
openai